chore(core): 更新缓存类文档和配置测试参数
- 将缓存类注释从通用缓存类更新为redis缓存类- 修复工具类中guid函数的随机种子类型转换问题
- 为统一收单服务类添加完整的文档注释
- 更新测试配置中的应用ID、密钥、商户ID和终端ID为实际测试参数

+++ src/Support/Tools.php
@@ -30,7 +30,7 @@ class Tools
      */
     public static function guid()
     {
-        mt_srand((double)microtime() * 10000);
+        mt_srand((int)microtime() * 10000);
         $charid = strtoupper(md5(uniqid(rand(), true)));
         $hyphen = chr(45);// "-"
         $uuid = chr(123)// "{"
